feat(swap): refactor token list modal
- separates component + styling logic into separate files

- renamed components to better reflect their purpose
